requirements.txt
asgiref==3.2.7
boto3==1.13.19
botocore==1.16.19
certifi==2019.9.11
chardet==3.0.4
Django==3.0.6
django-braces==1.14.0
django-crispy-forms==1.9.1
django-model-utils==4.0.0
django-star-ratings==0.9.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.10.0
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.9